# fuelflow/api/app.py
"""FastAPI localhost backend."""

from __future__ import annotations

import os
from contextlib import asynccontextmanager
from pathlib import Path
from typing import Any

from fastapi import FastAPI, Header, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field, field_validator

from fuelflow.engine.opt.locks import ConstraintParamLock, MoveLock, OptimizeLockContract

from fuelflow.services import (
    ServiceError,
    get_optimize_lock_capabilities,
    load_schedule_file,
    load_scenario_file,
    resolve_optimize_lock_effective,
    run_fork,
    run_optimization,
    run_simulation,
    save_schedule,
    save_scenario,
    startup_cleanup,
    validate_scenario_file,
)

WORKSPACE = Path(os.environ.get("OPTIFUEL_WORKSPACE", ".")).resolve()
ALLOW_NON_LOOPBACK = os.environ.get("OPTIFUEL_ALLOW_NON_LOOPBACK", "0") == "1"


@asynccontextmanager
async def lifespan(_: FastAPI):
    startup_cleanup(WORKSPACE)
    yield


app = FastAPI(title="OptiFuel API", version="0.1.0-alpha", lifespan=lifespan)

if ALLOW_NON_LOOPBACK:
    app.add_middleware(
        CORSMiddleware,
        allow_origins=["*"],
        allow_methods=["*"],
        allow_headers=["*"],
    )
else:
    app.add_middleware(
        CORSMiddleware,
        allow_origins=["http://127.0.0.1:5173", "http://localhost:5173"],
        allow_methods=["*"],
        allow_headers=["*"],
    )


class SaveRequest(BaseModel):
    data: dict[str, Any]
    expected_digest: str | None = Field(default=None, alias="if_match")

    model_config = {"populate_by_name": True}


class SimulateRequest(BaseModel):
    scenario_path: str
    schedule_path: str
    runtime_mode: str = "fail_fast"


class OptimizeRequest(BaseModel):
    scenario_path: str
    seed: int = 42
    time_limit_sec: float = 5.0
    seed_schedule_path: str | None = None
    lock_mode: str | None = None
    structure_mode: str | None = None
    move_locks: list[MoveLock] | None = None
    scenario_locks: list[str] | None = None
    constraint_param_locks: list[ConstraintParamLock] | None = None

    @field_validator("lock_mode")
    @classmethod
    def validate_lock_mode(cls, value: str | None) -> str | None:
        if value is None:
            return None
        if value not in {"legacy", "enforced"}:
            raise ValueError("lock_mode must be legacy or enforced")
        return value

    @field_validator("structure_mode")
    @classmethod
    def validate_structure_mode(cls, value: str | None) -> str | None:
        if value is None:
            return None
        if value not in {"locked", "unlocked"}:
            raise ValueError("structure_mode must be locked or unlocked")
        return value

    def to_lock_contract(self) -> OptimizeLockContract | None:
        if self.lock_mode is None:
            return None
        return OptimizeLockContract(
            lock_mode=self.lock_mode,  # type: ignore[arg-type]
            structure_mode=self.structure_mode,  # type: ignore[arg-type]
            move_locks=self.move_locks or [],
            scenario_locks=self.scenario_locks or [],
            constraint_param_locks=self.constraint_param_locks or [],
        )


class LockEffectiveRequest(BaseModel):
    scenario_path: str
    schedule_path: str | None = None
    lock_mode: str | None = None
    structure_mode: str | None = None
    constraint_param_locks: list[ConstraintParamLock] | None = None

    @field_validator("lock_mode")
    @classmethod
    def validate_lock_mode(cls, value: str | None) -> str | None:
        if value is None:
            return None
        if value not in {"legacy", "enforced"}:
            raise ValueError("lock_mode must be legacy or enforced")
        return value

    def to_lock_contract(self) -> OptimizeLockContract | None:
        if self.lock_mode is None:
            return None
        return OptimizeLockContract(
            lock_mode=self.lock_mode,  # type: ignore[arg-type]
            structure_mode=self.structure_mode,  # type: ignore[arg-type]
            constraint_param_locks=self.constraint_param_locks or [],
        )


class ForkRequest(BaseModel):
    scenario_path: str
    schedule_path: str
    state_at_min: float
    amendments: dict[str, Any]
    precedence: str | None = None
    output_path: str | None = None


def _handle_error(exc: Exception) -> None:
    if isinstance(exc, ServiceError):
        raise HTTPException(status_code=400, detail={"code": exc.code, "message": str(exc)}) from exc
    raise exc


@app.get("/health")
def health() -> dict[str, str]:
    return {"status": "ok", "release": "v1-alpha"}


@app.get("/scenarios")
def list_scenarios() -> list[str]:
    examples = WORKSPACE / "examples"
    if not examples.exists():
        return []
    return [str(p.relative_to(WORKSPACE)) for p in examples.rglob("scenario.yaml")]


@app.get("/scenarios/{scenario_id:path}")
def get_scenario(scenario_id: str) -> dict[str, Any]:
    path = WORKSPACE / scenario_id
    try:
        scenario, digest_value = load_scenario_file(path, WORKSPACE)
    except Exception as exc:
        _handle_error(exc)
    return {
        "data": scenario.model_dump(by_alias=True),
        "digest": digest_value,
        "etag": digest_value,
    }


@app.put("/scenarios/{scenario_id:path}")
def put_scenario(
    scenario_id: str,
    body: SaveRequest,
    if_match: str | None = Header(default=None),
) -> dict[str, str]:
    path = WORKSPACE / scenario_id
    try:
        digest_value = save_scenario(
            path,
            body.data,
            WORKSPACE,
            expected_digest=if_match or body.expected_digest,
        )
    except Exception as exc:
        _handle_error(exc)
    return {"digest": digest_value, "etag": digest_value}


@app.post("/scenarios/{scenario_id:path}/validate")
def validate_scenario_endpoint(scenario_id: str) -> dict[str, Any]:
    path = WORKSPACE / scenario_id
    try:
        return validate_scenario_file(path, WORKSPACE)
    except Exception as exc:
        _handle_error(exc)


@app.get("/schedules/{schedule_id:path}")
def get_schedule(schedule_id: str) -> dict[str, Any]:
    path = WORKSPACE / schedule_id
    try:
        schedule, digest_value = load_schedule_file(path, WORKSPACE)
    except Exception as exc:
        _handle_error(exc)
    return {
        "data": schedule.model_dump(by_alias=True),
        "digest": digest_value,
        "etag": digest_value,
    }


@app.put("/schedules/{schedule_id:path}")
def put_schedule(
    schedule_id: str,
    body: SaveRequest,
    if_match: str | None = Header(default=None),
) -> dict[str, str]:
    path = WORKSPACE / schedule_id
    try:
        digest_value = save_schedule(
            path,
            body.data,
            WORKSPACE,
            expected_digest=if_match or body.expected_digest,
        )
    except Exception as exc:
        _handle_error(exc)
    return {"digest": digest_value, "etag": digest_value}


@app.post("/runs/simulate")
def simulate_endpoint(body: SimulateRequest) -> dict[str, Any]:
    try:
        return run_simulation(
            WORKSPACE / body.scenario_path,
            WORKSPACE / body.schedule_path,
            WORKSPACE,
            runtime_mode=body.runtime_mode,
        )
    except Exception as exc:
        _handle_error(exc)


@app.get("/runs/optimize/capabilities")
def optimize_capabilities(
    scenario_path: str | None = None,
    schedule_path: str | None = None,
) -> dict[str, Any]:
    if scenario_path is None:
        from fuelflow.engine.opt.locks import lock_capabilities

        return lock_capabilities()
    try:
        return get_optimize_lock_capabilities(
            WORKSPACE / scenario_path,
            WORKSPACE,
            schedule_path=WORKSPACE / schedule_path if schedule_path else None,
        )
    except Exception as exc:
        _handle_error(exc)


@app.post("/runs/optimize/locks/effective")
def optimize_lock_effective(body: LockEffectiveRequest) -> dict[str, Any]:
    try:
        lock_contract = body.to_lock_contract()
        if lock_contract is None:
            lock_contract = OptimizeLockContract(lock_mode="enforced", structure_mode="locked")
        OptimizeLockContract.model_validate(lock_contract.model_dump())
        return resolve_optimize_lock_effective(
            WORKSPACE / body.scenario_path,
            WORKSPACE,
            schedule_path=WORKSPACE / body.schedule_path if body.schedule_path else None,
            lock_contract=lock_contract,
        )
    except ValueError as exc:
        raise HTTPException(
            status_code=400,
            detail={"code": "lock_contract_invalid", "message": str(exc)},
        ) from exc
    except Exception as exc:
        _handle_error(exc)


@app.post("/runs/optimize")
def optimize_endpoint(body: OptimizeRequest) -> dict[str, Any]:
    try:
        lock_contract = body.to_lock_contract()
        if lock_contract is not None:
            OptimizeLockContract.model_validate(lock_contract.model_dump())
        return run_optimization(
            WORKSPACE / body.scenario_path,
            WORKSPACE,
            seed=body.seed,
            time_limit_sec=body.time_limit_sec,
            seed_schedule_path=WORKSPACE / body.seed_schedule_path if body.seed_schedule_path else None,
            lock_contract=lock_contract,
        )
    except ValueError as exc:
        raise HTTPException(
            status_code=400,
            detail={"code": "lock_contract_invalid", "message": str(exc)},
        ) from exc
    except Exception as exc:
        _handle_error(exc)


@app.post("/scenarios/fork")
def fork_endpoint(body: ForkRequest) -> dict[str, Any]:
    try:
        return run_fork(
            WORKSPACE / body.scenario_path,
            WORKSPACE / body.schedule_path,
            WORKSPACE,
            state_at_min=body.state_at_min,
            amendments=body.amendments,
            precedence=body.precedence,
            output_path=WORKSPACE / body.output_path if body.output_path else None,
        )
    except Exception as exc:
        _handle_error(exc)