# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "ConstraintParamLock": {
        "properties": {
          "constraint_id": {
            "title": "Constraint Id",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "locked": {
            "default": true,
            "title": "Locked",
            "type": "boolean"
          }
        },
        "required": [
          "entity_id",
          "constraint_id"
        ],
        "title": "ConstraintParamLock",
        "type": "object"
      },
      "ForkRequest": {
        "properties": {
          "amendments": {
            "additionalProperties": true,
            "title": "Amendments",
            "type": "object"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "precedence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precedence"
          },
          "scenario_path": {
            "title": "Scenario Path",
            "type": "string"
          },
          "schedule_path": {
            "title": "Schedule Path",
            "type": "string"
          },
          "state_at_min": {
            "title": "State At Min",
            "type": "number"
          }
        },
        "required": [
          "scenario_path",
          "schedule_path",
          "state_at_min",
          "amendments"
        ],
        "title": "ForkRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LockEffectiveRequest": {
        "properties": {
          "constraint_param_locks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConstraintParamLock"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Param Locks"
          },
          "lock_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lock Mode"
          },
          "scenario_path": {
            "title": "Scenario Path",
            "type": "string"
          },
          "schedule_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Path"
          },
          "structure_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Mode"
          }
        },
        "required": [
          "scenario_path"
        ],
        "title": "LockEffectiveRequest",
        "type": "object"
      },
      "MoveLock": {
        "properties": {
          "edge": {
            "title": "Edge",
            "type": "string"
          },
          "entity": {
            "title": "Entity",
            "type": "string"
          },
          "locked_fields": {
            "items": {
              "const": "start_min",
              "type": "string"
            },
            "title": "Locked Fields",
            "type": "array"
          },
          "occurrence_index": {
            "default": 0,
            "title": "Occurrence Index",
            "type": "integer"
          }
        },
        "required": [
          "entity",
          "edge"
        ],
        "title": "MoveLock",
        "type": "object"
      },
      "OptimizeRequest": {
        "properties": {
          "constraint_param_locks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ConstraintParamLock"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Param Locks"
          },
          "lock_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lock Mode"
          },
          "move_locks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MoveLock"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Locks"
          },
          "scenario_locks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Locks"
          },
          "scenario_path": {
            "title": "Scenario Path",
            "type": "string"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          },
          "seed_schedule_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Schedule Path"
          },
          "structure_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Mode"
          },
          "time_limit_sec": {
            "default": 5.0,
            "title": "Time Limit Sec",
            "type": "number"
          }
        },
        "required": [
          "scenario_path"
        ],
        "title": "OptimizeRequest",
        "type": "object"
      },
      "SaveRequest": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "if_match": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "If Match"
          }
        },
        "required": [
          "data"
        ],
        "title": "SaveRequest",
        "type": "object"
      },
      "SimulateRequest": {
        "properties": {
          "runtime_mode": {
            "default": "fail_fast",
            "title": "Runtime Mode",
            "type": "string"
          },
          "scenario_path": {
            "title": "Scenario Path",
            "type": "string"
          },
          "schedule_path": {
            "title": "Schedule Path",
            "type": "string"
          }
        },
        "required": [
          "scenario_path",
          "schedule_path"
        ],
        "title": "SimulateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OptiFuel API",
    "version": "0.1.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/runs/optimize": {
      "post": {
        "operationId": "optimize_endpoint_runs_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Optimize Endpoint Runs Optimize Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Endpoint"
      }
    },
    "/runs/optimize/capabilities": {
      "get": {
        "operationId": "optimize_capabilities_runs_optimize_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "scenario_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Path"
            }
          },
          {
            "in": "query",
            "name": "schedule_path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Schedule Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Optimize Capabilities Runs Optimize Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Capabilities"
      }
    },
    "/runs/optimize/locks/effective": {
      "post": {
        "operationId": "optimize_lock_effective_runs_optimize_locks_effective_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockEffectiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Optimize Lock Effective Runs Optimize Locks Effective Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize Lock Effective"
      }
    },
    "/runs/simulate": {
      "post": {
        "operationId": "simulate_endpoint_runs_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Simulate Endpoint Runs Simulate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Endpoint"
      }
    },
    "/scenarios": {
      "get": {
        "operationId": "list_scenarios_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Scenarios Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios"
      }
    },
    "/scenarios/fork": {
      "post": {
        "operationId": "fork_endpoint_scenarios_fork_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fork Endpoint Scenarios Fork Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fork Endpoint"
      }
    },
    "/scenarios/{scenario_id}": {
      "get": {
        "operationId": "get_scenario_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Scenario Scenarios  Scenario Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario"
      },
      "put": {
        "operationId": "put_scenario_scenarios__scenario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Put Scenario Scenarios  Scenario Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Scenario"
      }
    },
    "/scenarios/{scenario_id}/validate": {
      "post": {
        "operationId": "validate_scenario_endpoint_scenarios__scenario_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Scenario Endpoint Scenarios  Scenario Id  Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Scenario Endpoint"
      }
    },
    "/schedules/{schedule_id}": {
      "get": {
        "operationId": "get_schedule_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Schedule Schedules  Schedule Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule"
      },
      "put": {
        "operationId": "put_schedule_schedules__schedule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Put Schedule Schedules  Schedule Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Schedule"
      }
    }
  }
}